Connects #1:  Make GitHub usernames case insensitive in the Team API.
This changeset makes the handling of GitHub usernames case insensitive when importing data into the Team API and adds several test cases to capture the scenarios in which this kind of problem can occur.  Ultimately we want it to be easy to just import an .about.yml file and not have to worry whether or not the GitHub usernames match.