Closes #57:  Remove Jekyll `site.collections` munging
This changeset removes the processing around the Jekyll `site.collections` hash since we are now dealing with the YAML files directly for the Team API.  This means that everything is already in the `site.data` hash instead.  Some additional processing was added to make sure the `site.data` hash has the correct keys.